Add PAADP
Physical Activity Accelerometer Data Processor (PAADP) is an open-source application for processing and summarizing raw physical activity accelerometer data into usable outputs. The application was designed using a custom Gtk3 front-end and R for the computational back-end. PAADP is Windows application. Versions for macOS and Linux are currently in development.